The saddest djs13 breaking change
djs no longer keeps a history of edits on a message. This commit removes the
functionality of displaying message edits on a message delete. I don't think it's
necessary to implement edit history ourselves because edits are logged as well,
but future work could include adding this functionality back in with a database
entry on message edit.